+++ src/Data display/Kanban/components/Item/Item.module.css
@@ -55,7 +55,6 @@
   display: flex;
   flex-grow: 1;
   align-items: center;
-  padding: 18px 20px;
   background-color: #fff;
   box-shadow: 0 0 0 calc(1px / var(--scale-x, 1)) rgba(63, 63, 68, 0.05),
     0 1px calc(3px / var(--scale-x, 1)) 0 rgba(34, 33, 81, 0.15);
