+++ src/Data display/Kanban_v1/Kanban_v1.stories.mdx
@@ -1,19 +1,19 @@
 import { Meta, Canvas, Props } from "@storybook/addon-docs/blocks";
 import { ArgsTable, Stories } from "@storybook/addon-docs";
-import KanbanV2 from "./index";
+import Kanban from "./index";
 import styled from "@emotion/styled";
-import Story from "./story";
+import Story from "./Story";
 
-<Meta title="Data Display/Kanban V2" component={KanbanV2} />
+<Meta title="Data Display/Kanban v1" component={Kanban} />
 
 export const TemplateDefault = (args) => {
   return <Story {...args} />;
 };
 
-# KanbanV2
+# Kanban
 
 ```jsx
-import { KanbanV2 } from "@lnc-react-ui";
+import { Kanban } from "@lnc-react-ui";
 ```
 
 <Canvas>
