--- src/Data display/Kanban_v2/story.js
@@ -1,81 +1,0 @@
-import React from "react";
-import styled from "@emotion/styled";
-import KanbanV2 from ".";
-import { useState, useCallback } from "react";
-import { arrayMove } from "@dnd-kit/sortable";
-import { cloneDeep } from "lodash";
-import { useEffect } from "react";
-
-const Container = styled.div``;
-
-const Story = (props) => {
-  const [data, setData] = useState({
-    A: [
-      { id: 1, name: "Card 1" },
-      { id: 2, name: "Card 2" },
-      { id: 3, name: "Card 3" },
-    ],
-    B: [
-      { id: 4, name: "Card 4" },
-      { id: 5, name: "Card 5" },
-      { id: 6, name: "Card 6" },
-    ],
-    C: [
-      { id: 7, name: "Card 7" },
-      { id: 8, name: "Card 8" },
-      { id: 9, name: "Card 9" },
-    ],
-  });
-
-  const onColumnMoved = (e, columns) => {
-    var newData = {};
-
-    columns.forEach((c) => {
-      newData[c] = data[c];
-    });
-
-    setData(cloneDeep(newData));
-  };
-
-  const onCardMoved = (e, items, column) => {
-    // var dataCopy = cloneDeep(data);
-    // dataCopy[column] = items[column];
-    // console.log(dataCopy);
-    // setData(dataCopy);
-  };
-
-  const onCardChangedColumns = (event, cachedItems, column, previousColumn) => {
-    // var dataCopy = cloneDeep(data);
-    // var current = dataCopy[column];
-    // dataCopy[column] = items[column];
-    // console.log(dataCopy);
-    // setData(dataCopy);
-  };
-
-  return (
-    <Container>
-      <KanbanV2
-        onColumnMoved={onColumnMoved}
-        onCardMoved={onCardMoved}
-        onCardChangedColumns={onCardChangedColumns}
-        data={data}
-        columnInfo={{
-          A: {
-            id: 1,
-            name: "Column A",
-          },
-          B: {
-            id: 2,
-            name: "Column B",
-          },
-          C: {
-            id: 3,
-            name: "Column C",
-          },
-        }}
-      />
-    </Container>
-  );
-};
-
-export default Story;
